Fix compile error of upgrading alluxio to 2.6.1
