--- presto-cache/src/main/java/com/facebook/presto/cache/alluxio/AlluxioURIStatus.java
@@ -1,37 +1,0 @@
-/*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.facebook.presto.cache.alluxio;
-
-import alluxio.client.file.URIStatus;
-import alluxio.wire.FileInfo;
-import com.facebook.presto.hive.HiveFileContext;
-
-import static java.util.Objects.requireNonNull;
-
-public class AlluxioURIStatus
-        extends URIStatus
-{
-    private final HiveFileContext hiveFileContext;
-
-    public AlluxioURIStatus(FileInfo info, HiveFileContext hiveFileContext)
-    {
-        super(info);
-        this.hiveFileContext = requireNonNull(hiveFileContext, "hiveFileContext is null");
-    }
-
-    public HiveFileContext getHiveFileContext()
-    {
-        return hiveFileContext;
-    }
-}
